Work on PR feedback
