+++ MiniTwit.Web.App/Models/Api/PostFollow.cs
@@ -1,0 +1,9 @@
+namespace MiniTwit.Web.App.Models.Api
+{
+    public class PostFollow
+    {
+        public string Follow { get; set; }
+        public string UnFollow { get; set; }
+        public int? Latest { get; set; }
+    }
+}
