+++ MiniTwit.Web.App/Models/Api/PostMessage.cs
@@ -1,0 +1,8 @@
+namespace MiniTwit.Web.App.Models.Api
+{
+    public class MessagePost
+    {
+        public string Content { get; set; }
+        public int? Latest { get; set; }
+    }
+}
